+++ core/src/com/biglybt/core/networkmanager/RateHandler.java
@@ -27,11 +27,11 @@ public interface RateHandler {
    * Get the current number of bytes allowed to be processed by the entity and protocol_is_free if [1] > 0
    * @return number of bytes allowed
    */
-  public int[] getCurrentNumBytesAllowed();
+  public long[] getCurrentNumBytesAllowed();
 
   /**
    * Notification of any bytes processed by the entity.
    * @param num_bytes_processed
    */
-  public void bytesProcessed( int data_bytes, int protocol_bytes );
+  public void bytesProcessed( long data_bytes, long protocol_bytes );
 }
