+++ core/src/com/biglybt/core/networkmanager/impl/SinglePeerUploader.java
@@ -75,7 +75,7 @@ public class SinglePeerUploader implements RateControlledEntity {
 			if( connection.getOutgoingMessageQueue().getTotalSize() < 1 ) {
 				return false;  //no data to send
 			}
-			int[] allowed = rate_handler.getCurrentNumBytesAllowed();
+			long[] allowed = rate_handler.getCurrentNumBytesAllowed();
 
 			if( allowed[0] < 1 ){
 
@@ -109,7 +109,7 @@ public class SinglePeerUploader implements RateControlledEntity {
 	}
 
 	@Override
-	public int 
+	public long 
 	doProcessing(
 		EventWaiter waiter, 
 		int max_bytes ) 
@@ -122,9 +122,9 @@ public class SinglePeerUploader implements RateControlledEntity {
 				return 0;
 			}
 	
-			int[] allowed = rate_handler.getCurrentNumBytesAllowed();
+			long[] allowed = rate_handler.getCurrentNumBytesAllowed();
 	
-			int num_bytes_allowed = allowed[0];
+			long num_bytes_allowed = allowed[0];
 	
 			boolean	protocol_is_free = allowed[1] > 0;
 	
@@ -157,7 +157,7 @@ public class SinglePeerUploader implements RateControlledEntity {
 				return 0;
 			}
 	
-			int num_bytes_to_write = num_bytes_allowed > num_bytes_available ? num_bytes_available : num_bytes_allowed;
+			long num_bytes_to_write = num_bytes_allowed > num_bytes_available ? num_bytes_available : num_bytes_allowed;
 	
 			//int mss = NetworkManager.getTcpMssSize();
 			//if( num_bytes_to_write > mss )  num_bytes_to_write = mss;
@@ -265,11 +265,11 @@ public class SinglePeerUploader implements RateControlledEntity {
   public String
   getString()
   {
-	  int[] temp = rate_handler.getCurrentNumBytesAllowed();
+	  long[] temp = rate_handler.getCurrentNumBytesAllowed();
 	  
 	  String ba = "";
 	  
-	  for ( int t: temp ){
+	  for ( long t: temp ){
 		  ba += (ba.isEmpty()?"":",") + t;
 	  }
