+++ core/src/com/biglybt/core/peermanager/messaging/azureus/AZMessageDecoder.java
@@ -83,11 +83,11 @@ public class AZMessageDecoder implements MessageStreamDecoder {
   }
 
   @Override
-  public int performStreamDecode(Transport transport, int max_bytes ) throws IOException {
+  public long performStreamDecode(Transport transport, long max_bytes ) throws IOException {
     protocol_bytes_last_read = 0;
     data_bytes_last_read = 0;
 
-    int bytes_remaining = max_bytes;
+    long bytes_remaining = max_bytes;
 
     while( bytes_remaining > 0 ) {
       if( destroyed ) {
@@ -102,7 +102,7 @@ public class AZMessageDecoder implements MessageStreamDecoder {
         break;
       }
 
-      int bytes_possible = preReadProcess( bytes_remaining );
+      long bytes_possible = preReadProcess( bytes_remaining );
 
       if( bytes_possible < 1 ) {
         Debug.out( "ERROR AZ: bytes_possible < 1" );
@@ -223,7 +223,7 @@ public class AZMessageDecoder implements MessageStreamDecoder {
 
 
 
-  private int preReadProcess( int allowed ) {
+  private long preReadProcess( long allowed ) {
     if( allowed < 1 ) {
       Debug.out( "allowed < 1" );
     }
@@ -260,7 +260,7 @@ public class AZMessageDecoder implements MessageStreamDecoder {
         }
 
         if( remaining > allowed ) {  //read only part of this buffer
-          bb.limit( bb.position() + allowed );  //limit current buffer
+          bb.limit((int)( bb.position() + allowed ));  //limit current buffer
           bytes_available += bb.remaining();
           shrink_remaining_buffers = true;  //shrink any tail buffers
         }
